Moving phrase types to a separate file
