+++ src/types/font.ts
@@ -1,3 +1,4 @@
+// TODO: rename Element as it's reserved name
 export interface Element {
 	elementDelay: number;
 	shape: string;
